refactor: switch E2E tests to data plane signaling
