+++ system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/edc/test/e2e/InMemorySignalingRuntimes.java
@@ -12,16 +12,15 @@
  *
  */
 
-package org.eclipse.edc.test.e2e.signaling;
+package org.eclipse.edc.test.e2e;
 
 import org.eclipse.edc.junit.extensions.EdcClassRuntimesExtension;
 import org.eclipse.edc.junit.extensions.EdcRuntimeExtension;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import java.util.HashMap;
-
-import static org.eclipse.edc.test.e2e.signaling.SignalingEndToEndTestBase.CONSUMER;
-import static org.eclipse.edc.test.e2e.signaling.SignalingEndToEndTestBase.PROVIDER;
+import static org.eclipse.edc.test.e2e.Runtimes.backendService;
+import static org.eclipse.edc.test.e2e.TransferEndToEndTestBase.CONSUMER;
+import static org.eclipse.edc.test.e2e.TransferEndToEndTestBase.PROVIDER;
 
 public interface InMemorySignalingRuntimes {
 
@@ -54,30 +53,14 @@ public interface InMemorySignalingRuntimes {
                     CONSUMER.controlPlaneConfiguration(),
                     CONTROL_PLANE_MODULES
             ),
-            new EdcRuntimeExtension(
-                    ":system-tests:e2e-transfer-test:backend-service",
-                    "consumer-backend-service",
-                    new HashMap<>() {
-                        {
-                            put("web.http.port", String.valueOf(CONSUMER.backendService().getPort()));
-                        }
-                    }
-            ),
+            backendService("consumer-backend-service", CONSUMER.backendServiceConfiguration()),
             DATA_PLANE,
             new EdcRuntimeExtension(
                     "provider-control-plane",
                     PROVIDER.controlPlaneConfiguration(),
                     CONTROL_PLANE_MODULES
             ),
-            new EdcRuntimeExtension(
-                    ":system-tests:e2e-transfer-test:backend-service",
-                    "provider-backend-service",
-                    new HashMap<>() {
-                        {
-                            put("web.http.port", String.valueOf(PROVIDER.backendService().getPort()));
-                        }
-                    }
-            )
+            backendService("provider-backend-service", PROVIDER.backendServiceConfiguration())
     );
