+++ system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/edc/test/e2e/TransferEndToEndTestBase.java
@@ -15,28 +15,37 @@
 package org.eclipse.edc.test.e2e;
 
 import jakarta.json.JsonObject;
-import org.eclipse.edc.test.e2e.participant.EndToEndTransferParticipant;
+import org.eclipse.edc.test.e2e.participant.TransferEndToEndParticipant;
 
 import java.time.Duration;
 import java.util.Map;
+import java.util.Set;
 import java.util.UUID;
+
+import static org.eclipse.edc.connector.controlplane.test.system.utils.PolicyFixtures.noConstraintPolicy;
 
 public abstract class TransferEndToEndTestBase {
 
     protected final Duration timeout = Duration.ofSeconds(60);
 
-    protected static final EndToEndTransferParticipant CONSUMER = EndToEndTransferParticipant.Builder.newInstance()
+    protected static final TransferEndToEndParticipant CONSUMER = TransferEndToEndParticipant.Builder.newInstance()
             .name("consumer")
             .id("urn:connector:consumer")
             .build();
-    protected static final EndToEndTransferParticipant PROVIDER = EndToEndTransferParticipant.Builder.newInstance()
+    protected static final TransferEndToEndParticipant PROVIDER = TransferEndToEndParticipant.Builder.newInstance()
             .name("provider")
             .id("urn:connector:provider")
             .build();
 
-    protected void createResourcesOnProvider(String assetId, JsonObject policy, Map<String, Object> dataAddressProperties) {
+    protected void createResourcesOnProvider(String assetId, JsonObject contractPolicy, Map<String, Object> dataAddressProperties) {
         PROVIDER.createAsset(assetId, Map.of("description", "description"), dataAddressProperties);
-        var policyDefinition = PROVIDER.createPolicyDefinition(policy);
-        PROVIDER.createContractDefinition(assetId, UUID.randomUUID().toString(), policyDefinition, policyDefinition);
+        var accessPolicyId = PROVIDER.createPolicyDefinition(noConstraintPolicy());
+        var contractPolicyId = PROVIDER.createPolicyDefinition(contractPolicy);
+        PROVIDER.createContractDefinition(assetId, UUID.randomUUID().toString(), accessPolicyId, contractPolicyId);
     }
+
+    protected void registerDataPlanes() {
+        PROVIDER.registerDataPlane(Set.of("HttpData-PUSH", "HttpData-PULL"));
+    }
+
 }
