+++ system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/edc/test/e2e/TransferSignalingPushEndToEndTest.java
@@ -12,17 +12,19 @@
  *
  */
 
-package org.eclipse.edc.test.e2e.signaling;
+package org.eclipse.edc.test.e2e;
 
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
 import org.eclipse.edc.junit.annotations.EndToEndTest;
+import org.eclipse.edc.junit.annotations.PostgresqlIntegrationTest;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.BeforeAllCallback;
+import org.junit.jupiter.api.extension.RegisterExtension;
 
-import java.time.Duration;
 import java.util.Map;
 import java.util.UUID;
 
@@ -34,6 +36,7 @@ import static org.eclipse.edc.connector.controlplane.test.system.utils.PolicyFix
 import static org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferProcessStates.COMPLETED;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.TYPE;
 import static org.eclipse.edc.spi.constants.CoreConstants.EDC_NAMESPACE;
+import static org.eclipse.edc.sql.testfixtures.PostgresqlEndToEndInstance.createDatabase;
 import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
@@ -53,8 +56,18 @@ class TransferSignalingPushEndToEndTest {
 
     }
 
-    abstract static class Tests extends SignalingEndToEndTestBase {
-        protected final Duration timeout = Duration.ofSeconds(60);
+    @Nested
+    @PostgresqlIntegrationTest
+    class Postgres extends Tests implements PostgresSignalingRuntimes {
+
+        @RegisterExtension
+        static final BeforeAllCallback CREATE_DATABASES = context -> {
+            createDatabase(CONSUMER.getName());
+            createDatabase(PROVIDER.getName());
+        };
+    }
+
+    abstract static class Tests extends TransferEndToEndTestBase {
 
         @BeforeEach
         void beforeEach() {
@@ -68,6 +81,36 @@ class TransferSignalingPushEndToEndTest {
             var destination = httpDataAddress(CONSUMER.backendService() + "/api/consumer/store");
 
             var transferProcessId = CONSUMER.requestAsset(PROVIDER, assetId, noPrivateProperty(), destination, "HttpData-PUSH");
+            await().atMost(timeout).untilAsserted(() -> {
+                var state = CONSUMER.getTransferProcessState(transferProcessId);
+                assertThat(state).isEqualTo(COMPLETED.name());
+
+                given()
+                        .baseUri(CONSUMER.backendService().toString())
+                        .when()
+                        .get("/api/consumer/data")
+                        .then()
+                        .statusCode(anyOf(is(200), is(204)))
+                        .body(is(notNullValue()));
+            });
+        }
+
+        @Test
+        void httpToHttp_oauth2Provisioning() {
+            var assetId = UUID.randomUUID().toString();
+            var sourceDataAddressProperties = Map.<String, Object>of(
+                    "type", "HttpData",
+                    "baseUrl", PROVIDER.backendService() + "/api/provider/oauth2data",
+                    "oauth2:clientId", "clientId",
+                    "oauth2:clientSecretKey", "provision-oauth-secret",
+                    "oauth2:tokenUrl", PROVIDER.backendService() + "/api/oauth2/token"
+            );
+
+            createResourcesOnProvider(assetId, noConstraintPolicy(), sourceDataAddressProperties);
+            var destination = httpDataAddress(CONSUMER.backendService() + "/api/consumer/store");
+
+            var transferProcessId = CONSUMER.requestAsset(PROVIDER, assetId, noPrivateProperty(), destination, "HttpData-PUSH");
+
             await().atMost(timeout).untilAsserted(() -> {
                 var state = CONSUMER.getTransferProcessState(transferProcessId);
                 assertThat(state).isEqualTo(COMPLETED.name());
