--- system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/edc/test/e2e/participant/EndToEndTransferParticipant.java
@@ -1,44 +1,0 @@
-/*
- *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.test.e2e.participant;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-
-import java.net.URI;
-
-import static org.eclipse.edc.util.io.Ports.getFreePort;
-
-public class EndToEndTransferParticipant extends BaseEndToEndParticipant {
-
-    public static final class Builder extends BaseEndToEndParticipant.Builder<EndToEndTransferParticipant, Builder> {
-
-        private Builder() {
-            super(new EndToEndTransferParticipant());
-        }
-
-        @JsonCreator
-        public static Builder newInstance() {
-            return new Builder();
-        }
-
-        @Override
-        public EndToEndTransferParticipant build() {
-            super.managementEndpoint(new Endpoint(URI.create("http://localhost:" + getFreePort() + "/api/management")));
-            super.protocolEndpoint(new Endpoint(URI.create("http://localhost:" + getFreePort() + "/protocol")));
-            super.build();
-            return participant;
-        }
-    }
-}
