--- system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/edc/test/e2e/participant/SignalingParticipant.java
@@ -1,92 +1,0 @@
-/*
- *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.test.e2e.participant;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.restassured.common.mapper.TypeRef;
-import org.eclipse.edc.spi.types.domain.DataAddress;
-import org.hamcrest.Matcher;
-
-import java.util.Map;
-
-import static io.restassured.RestAssured.given;
-import static io.restassured.http.ContentType.JSON;
-
-public class SignalingParticipant extends BaseEndToEndParticipant {
-
-    /**
-     * Get the EDR from the EDR cache by transfer process id.
-     *
-     * @param transferProcessId The transfer process id
-     * @return The cached {@link DataAddress}
-     */
-    public DataAddress getEdr(String transferProcessId) {
-        var dataAddressRaw = managementEndpoint.baseRequest()
-                .contentType(JSON)
-                .when()
-                .get("/v1/edrs/{id}/dataaddress", transferProcessId)
-                .then()
-                .log().ifError()
-                .statusCode(200)
-                .contentType(JSON)
-                .extract().body().as(new TypeRef<Map<String, Object>>() {
-                });
-
-
-        var builder = DataAddress.Builder.newInstance();
-        dataAddressRaw.forEach(builder::property);
-        return builder.build();
-
-    }
-
-    /**
-     * Pull data from provider using EDR.
-     *
-     * @param edr         endpoint data reference
-     * @param queryParams query parameters
-     * @param bodyMatcher matcher for response body
-     */
-    public void pullData(DataAddress edr, Map<String, String> queryParams, Matcher<String> bodyMatcher) {
-        given()
-                .baseUri(edr.getStringProperty("endpoint"))
-                .header("Authorization", edr.getStringProperty("authorization"))
-                .queryParams(queryParams)
-                .when()
-                .get()
-                .then()
-                .log().ifError()
-                .statusCode(200)
-                .body("message", bodyMatcher);
-    }
-
-    public static class Builder extends BaseEndToEndParticipant.Builder<SignalingParticipant, Builder> {
-
-        protected Builder() {
-            super(new SignalingParticipant());
-        }
-
-        @JsonCreator
-        public static Builder newInstance() {
-            return new Builder();
-        }
-
-        @Override
-        public SignalingParticipant build() {
-            super.build();
-            return participant;
-        }
-
-    }
-}
